requirements.txt
appdirs==1.4.3
asgiref==3.2.3
attrs==19.3.0
autobahn==20.2.1
Automat==20.2.0
black==19.10b0
boto3==1.12.3
botocore==1.15.3
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
daphne==2.4.1
Django==3.0.3
django-cors-headers==3.2.1
django-dotenv==1.4.2
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
drf-yasg==1.17.1
entrypoints==0.3
flake8==3.7.9
flex==6.14.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
inflection==0.3.1
itypes==1.1.0
Jinja2==3.0.0a1
jmespath==0.9.4
jsonpointer==2.0
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.1
pathspec==0.7.0
Pillow==7.0.0
psycopg2-binary==2.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
PyHamcrest==2.0.0
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
regex==2020.1.8
requests==2.23.0
rfc3987==1.3.8
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.0
strict-rfc3339==0.7
toml==0.10.0
Twisted==19.10.0
txaio==20.1.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
validate-email==1.3
zope.interface==4.7.1
